feat : 포인트 사용 기능 구현 및 DTO 리팩토링(패키지 구분화)
