--- src/main/java/org/example/fitpass/domain/point/dto/PointRequestDto.java
@@ -1,5 +1,0 @@
-package org.example.fitpass.domain.point.dto;
-
-public class PointRequestDto {
-
-}
