+++ src/main/java/org/example/fitpass/domain/point/dto/request/PointRequestDto.java
@@ -1,0 +1,5 @@
+package org.example.fitpass.domain.point.dto.request;
+
+public class PointRequestDto {
+
+}
