+++ src/main/java/org/example/fitpass/domain/point/dto/response/PointResponseDto.java
@@ -1,4 +1,4 @@
-package org.example.fitpass.domain.point.dto;
+package org.example.fitpass.domain.point.dto.response;
 
 import java.time.LocalDateTime;
 import lombok.Builder;
