+++ src/main/java/org/example/fitpass/domain/reservation/controller/ReservationController.java
@@ -6,13 +6,13 @@ import java.time.LocalTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.example.fitpass.common.ResponseMessage;
-import org.example.fitpass.domain.reservation.dto.GetReservationResponseDto;
-import org.example.fitpass.domain.reservation.dto.ReservationRequestDto;
-import org.example.fitpass.domain.reservation.dto.ReservationResponseDto;
-import org.example.fitpass.domain.reservation.dto.TrainerReservationResponseDto;
-import org.example.fitpass.domain.reservation.dto.UpdateReservationRequestDto;
-import org.example.fitpass.domain.reservation.dto.UpdateReservationResponseDto;
-import org.example.fitpass.domain.reservation.dto.UserReservationResponseDto;
+import org.example.fitpass.domain.reservation.dto.response.GetReservationResponseDto;
+import org.example.fitpass.domain.reservation.dto.request.ReservationRequestDto;
+import org.example.fitpass.domain.reservation.dto.response.ReservationResponseDto;
+import org.example.fitpass.domain.reservation.dto.response.TrainerReservationResponseDto;
+import org.example.fitpass.domain.reservation.dto.request.UpdateReservationRequestDto;
+import org.example.fitpass.domain.reservation.dto.response.UpdateReservationResponseDto;
+import org.example.fitpass.domain.reservation.dto.response.UserReservationResponseDto;
 import org.example.fitpass.domain.reservation.service.ReservationService;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
