+++ src/main/java/org/example/fitpass/domain/reservation/service/ReservationService.java
@@ -9,13 +9,13 @@ import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.example.fitpass.domain.gym.entity.Gym;
 import org.example.fitpass.domain.gym.repository.GymRepository;
-import org.example.fitpass.domain.reservation.dto.GetReservationResponseDto;
-import org.example.fitpass.domain.reservation.dto.ReservationRequestDto;
-import org.example.fitpass.domain.reservation.dto.ReservationResponseDto;
-import org.example.fitpass.domain.reservation.dto.TrainerReservationResponseDto;
-import org.example.fitpass.domain.reservation.dto.UpdateReservationRequestDto;
-import org.example.fitpass.domain.reservation.dto.UpdateReservationResponseDto;
-import org.example.fitpass.domain.reservation.dto.UserReservationResponseDto;
+import org.example.fitpass.domain.reservation.dto.response.GetReservationResponseDto;
+import org.example.fitpass.domain.reservation.dto.request.ReservationRequestDto;
+import org.example.fitpass.domain.reservation.dto.response.ReservationResponseDto;
+import org.example.fitpass.domain.reservation.dto.response.TrainerReservationResponseDto;
+import org.example.fitpass.domain.reservation.dto.request.UpdateReservationRequestDto;
+import org.example.fitpass.domain.reservation.dto.response.UpdateReservationResponseDto;
+import org.example.fitpass.domain.reservation.dto.response.UserReservationResponseDto;
 import org.example.fitpass.domain.reservation.entity.Reservation;
 import org.example.fitpass.domain.reservation.enums.ReservationStatus;
 import org.example.fitpass.domain.reservation.repository.ReservationRepository;
